refactor(ark): extract shared MCP client into internal/mcp/ package
- Move MCPClient, MCPSettings, MCPClientPool, BuildMCPServerURL from genai to internal/mcp
- MCPExecutor stays in genai (depends on execution types)
- Update controller and genai imports
